fix(math envs): Only change environments not allowed in math. (#308)
Partially fixes #298.
Environments like `align` are not allowedin latex math environments
and pandoc doesn't take those out of the math environments thus a filter was made to partially fix that.
But that filter captures too much and shouldn't capture environments like matrix and aligned and more.
Thus for <3.8 only those environments are fixed
for 3.8-3.8.2.x space is deleted because of https://github.com/jgm/pandoc/issues/11266.
And for >=3.8.3 that fix should be included thus no extra parsing is needed.